+++ .github/workflows/kind.yml
@@ -740,7 +740,7 @@ jobs:
           retention-days: 30
 
   conduct-connectivity-test:
-    name: Test connectivity using antctl test command
+    name: Test connectivity using 'antctl check' command
     needs: [ build-antrea-coverage-image ]
     runs-on: [ ubuntu-latest ]
     steps:
@@ -770,14 +770,14 @@ jobs:
       - name: Create Kind Cluster
         run: |
           kind create cluster --config ci/kind/config-3nodes.yml
-      - name: load docker image into kind
+      - name: Load docker image into kind
         run: |
           kind load docker-image antrea/antrea-controller-ubuntu-coverage:latest antrea/antrea-agent-ubuntu-coverage:latest
           kubectl apply -f build/yamls/antrea.yml
-      - name: build antctl binaries
+      - name: Build antctl binary
         run: |
           make antctl-linux
-      - name: run antctl command
+      - name: Run antctl command
         run: |
           ./bin/antctl-linux check installation
